Hold a float in an int register as its bits, not the other way about
The eight shaders the instruction writer computed wrongly were all one
cause. fxc does not read asfloat as a reinterpretation: it reads it as
producing a float and flushes a denormal to zero, so the bits of any
small integer put through one come back as zero. A register the writer
kept as a float, whose bits an integer instruction read, lost them - a
mask of 1, a sign bit, an index anded with 63, every step of fxc's own
f32tof16.

So the storage is inverted. A register any of whose components ever
holds bits is declared int, and the floats in it are held as their
bits: asfloat to read one, asint to write one. That direction fxc keeps
faith with, since a float's bits through an asfloat are the float, a
number it has no reason to flush. ComponentStorage loses its third
state along with the machinery that served it: a temp register is now
an int one or a float one, and which way a value crosses between them
is decided by the value, not by the component.

What a component of an int register holds is a float where a float
instruction writes it, reads it, or an itof converted it, or where a
mov carried one in - a mov says nothing about what it carries, so what
it came from has to, followed to a fixpoint through chains of them. An
integer the shader takes apart bitwise is not one of these: bits are
what an int variable holds anyway. A loop counter sharing the register
with both is a plain number and stays one.

Four things had to be got right around the edges. A conversion out of
an int register drops its `(int)` cast - the integer is there already -
but keeps `(uint)`, which says which half of the range the bits mean.
A negation or an absolute on bits a mov carries is a float operation:
`movc r2, r3, r2, -r2` over a fraction negates the number, and negating
its bits is a different number. A move between registers of one
declaration carries what is there, bits and all, and only a move across
converts - reading two sets of bits as floats to select between them is
the flush again. And a float immediate moved into an int register is
written as the integer its bits are: `asint(1)` is not them, a whole
float printing without a decimal point.

Three more came out of the same pass. An indexable temp declared int
reports its storage as one rather than defaulting to float. A load from
groupshared memory is typed by the array's own type, which the stores
into it settle, where the reflection data has nothing to say. And a
saturating move carries a float, [0, 1] being nothing an integer is put
through.

KnownDifferences is empty for the first time since it was added. 1649
tests against fxc under Wine, 0 failed, 0 skipped: the instruction
writer computes what every original does, as the AST writer already
did.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/int_constant_divide.fx
@@ -14,11 +14,11 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 	r0.y = (uint)r0.y % 13;
 	r0.w = -r0.z;
 	r0.x = (r0.x != 0) ? r0.w : r0.z;
-	r1.x = (float)(int)r0.x;
+	r1.x = (float)r0.x;
 	r0.x = -r0.y;
 	r0.z = value & -2147483648;
 	r0.x = (r0.z != 0) ? r0.x : r0.y;
-	r1.y = (float)(int)r0.x;
+	r1.y = (float)r0.x;
 	r0.x = (uint)uvalue / 10;
 	r1.z = (float)(uint)r0.x;
 	r0.x = (uint)uvalue % 6;
